mfg deleted alles
