--- src/mod_downloader_with_config/config.json
@@ -1,12 +1,0 @@
-{
-    "version": "1.21",
-    "loader": "fabric",
-    "mods": [
-        "sodium",
-        "lithium",
-        "fabric-api",
-        "krypton",
-        "optifine",
-        "zoomify"
-    ]
-}
